Fix test --- - Modify test data structure   - Add data for iterable for column (this should be failure at the time) - Split unit/functional test case out into another class
